net/iavf: fix vector id assignment
When the interrupt handle instance allows none packet I/O interrupts,
the max value of vector_id is set to be less than the number of msix
interrupts, but not equal to it. Which may cause same vector_id is sent
to PF with opcode VIRTCHNL_OP_CONFIG_IRQ_MAP to setup the cause of
interrupt mapping in iavf_config_irq_map(), these are invalid
parameters, cause error: VIRTCHNL_STATUS_ERR_PARAM(-5). So add code to
set the max value of vector_id to be equal to the number of msix
interrupts.

Fixes: d6bde6b5eae9 ("net/avf: enable Rx interrupt")
Cc: stable@dpdk.org

Signed-off-by: YU DAPENG <dapengx.yu@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>