net/iavf: fix vector id assignment
The number of MSI-X interrupts on Rx shall be the minimal value of the
number of available MSI-X interrupts per VF - 1 (the 1 is for
miscellaneous interrupt) and the number of configured Rx queues.
The current code break the rule because the number of available MSI-X
interrupts is used as the first value, but code does not subtract 1
from it.

In normal situation, the first value is larger than the second value. So
each queue can be assigned a unique vector_id.

For example: 17 available MSI-X interrupts, and 16 available Rx queues
per VF; but only 4 Rx queues are configured when device is started.
vector_id:0 is for misc interrupt, vector_id:1 for Rx queue0,
vector_id:2 for Rx queue1, vector_id:3 for Rx queue2, vector_id:4 for Rx
queue3.

Current code breaks the rule in this normal situation, because when
assign vector_ids to interrupt handle, for example, it does not assign
vector_id:4 to the queue3, but assign vector_id:1 to it, because the
condition used causes vector_id wrap around too early.

In iavf_config_irq_map(), the current code does not write data into the
last element of vecmap[], because of the previous code break. Which
cause wrong data is sent to PF with opcode VIRTCHNL_OP_CONFIG_IRQ_MAP
and cause error: VIRTCHNL_STATUS_ERR_PARAM(-5).

If kernel driver supports large VFs (up to 256 queues), different queues
can be assigned same vector_id.

In order to adapt to large VFs and avoid wrapping early, the condition
is replaced from vec >= vf->nb_msix to vec >= vf->vf_res->max_vectors.

Fixes: d6bde6b5eae9 ("net/avf: enable Rx interrupt")
Cc: stable@dpdk.org

Signed-off-by: YU DAPENG <dapengx.yu@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>