app/testpmd: add support for generic copy rte flow action
Add support for the RTE_FLOW_ACTION_COPY_FIELD to the testpmd.
Implement CLI to create the copy_field action and supply all the
needed parameters to copy an arbitrary packet field (as well as
mark, tag or metadata) into another field.

Example of the flow is the following:
flow create 0 egress group 1 pattern eth / ipv4 / udp / end
     actions copy_field dst_type tag dst_level 0 dst_offset 8
     src_type gtp_teid src_level 0 src_offset 0 width 32 / end

This flow copies 32 bits from the first Tag in the Tags array
into the outermost GTP TEID packet header field. 8 bits of the
Tag are skipped as indicated by the dst_offset action parameter.

dst_type and src_type are the only mandatory parameters to
specify. Levels and offset are 0 by default if they are not
overridden by a user. The width parameter gets the smallest width
from the source and destination sizes if it is not specified.

Signed-off-by: Alexander Kozyrev <akozyrev@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>