test: don't break terminal settings when running tests
When running one test (via DPDK_TEST) the test program
would leave the terminal in raw mode.  This was because
it was setting up cmdline to do interactive input.

The fix is to use cmdline_new() for the interactive
case.

This also fixes a memory leak because the test
runner was never calling cmdline_free().

Fixes: 9b848774a5dc ("test: use env variable to run tests")
Cc: harry.van.haaren@intel.com
Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>

revise test bugfix

Signed-off-by: 0-day Robot <robot@bytheb.org>