net/ixgbe: fix configuration of max frame size
For 82599 NIC, jumbo frame feature is not supported in IOV mode,
if a VF requests to configure the max frame size to that not
bigger than RTE_ETHER_MAX_LEN, the kernel driver returns 0, but
the DPDK ixgbe PMD returens -1, this will cause the VF to fail
to start when the PF driven by DPDK ixgbe PMD.

This patch keeps ixgbe PMD's handling mode consistent with kernel
driver in above situation.

In addition, the value set by the command IXGBE_VF_SET_LPE
represents the max frame size, not the mtu.

Fixes: 1b9ea09c067b ("ixgbe: support X550")
Fixes: 95a27b3ba5f5 ("net/ixgbe: enable jumbo frame for VF")
Cc: stable@dpdk.org

Signed-off-by: Alvin Zhang <alvinx.zhang@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>