net/octeontx: fix the jumbo frame flag condition for mtu set
The jumbo frame uses the 'RTE_ETHER_MAX_LEN' as boundary condition,
but the Ether overhead is larger than 18 when it supports dual VLAN tags.
That will cause the jumbo flag rx offload is wrong when MTU size is
'RTE_ETHER_MTU'.

This fix will change the boundary condition with 'RTE_ETHER_MTU' and
overhead, that perhaps impacts the cases of the jumbo frame related.

Fixes: 3151e6a687a3 ("net/octeontx: support MTU")

Cc: Harman Kalra <hkalra@marvell.com>

Acked-by: Harman Kalra <hkalra@marvell.com>

Signed-off-by: Steve Yang <stevex.yang@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>