net/ice: drain out DCF AdminQ command queue
The virtchnl command message response is handled by matchiing the opcode
only for limitation.

The DCF AdminQ command with buffer data needs two virtchnl commands, one
is to handle the AdminQ header, the other is to handle AdminQ buffer. If
the AdminQ header command gets the failure response, the AdminQ buffer
command needs to wait for the buffer message response until timeout to
drain out the virtchnl command queue, since both of them are sent to PF,
the PF will handle them one by one, and send back the response. If not,
it will cause the next AdminQ command failure with the stall response.

Fixes: daa714d55c72 ("net/ice: handle AdminQ command by DCF")
Cc: stable@dpdk.org

Signed-off-by: Haiyue Wang <haiyue.wang@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>